BUG: Combine adjacent whitespace delimiters in LYNX Orchestra loader
Because the CSV delimiter for LYNX Orchestra data is space, multiple
spaces in a row would cause an empty entry to be created in a the
row.

Adding the skipinitialspace option to the CSV reader constructor
causes these extra spaces to be aborbed into the nearest delimiter
instead of spawning an extra element.